requirements.txt
streamlit>=1.30.0
numpy
tensorflow>=2.18.0
scikit-learn
Pillow
joblib

